Added output from make to the README.
When the README described the action of the default make command it
just showed make being invoked. That was not very exciting or
particularly useful. So instead it was replaced by the output of the
command which gives a description of all the targets available with
a short description of each.